deploy: Use `cygpath` to translate `ldd` DLL paths
This patch adds a step in the get_deps.py script to translate the
UNIX style file paths given by `ldd`, which outputs paths of the
required DLLs for XQEMU, to the Windows style equivalents for
deployment of XQEMU on the Windows platform.

This step is necessary due to differences in path handling on
variants of Python in the MSYS2 environment. All variants seem
to handle the Windows style, but some do not understand the UNIX
style.