fpu: Begin using host FPU for hard x87 emulation
This is a start to hard float support for x87. It can fail in several
ways and has lots of room for improvement, but based on practical
testing it appears reliable and significantly increases performance in
FP-bottlenecked game code. Improvement to rounding, exception handling,
and performance accompanied by rigorous testing will be necessary next
steps for this work.