Fix for all scale factors
Since screen_pos (and therefore also pixel) are multiplied by the scale factor but the correct subpixel offset is always the same, it was fair to assume that the final subpixel value had to be divided by the scale factor for it to be corrected.

Tested on pgraph tests, Fable: TLC and Halo: CE (weapon zoom in).