nv2a: Update Conexant field counter on vblank
DreamWorks: Over the Hedge performs D3D resets before playing FMVs at startup.
As a part of this process, an initialization is performed that checks to see if
a DirectX frame counter aligns with the counter reported by the TV encoder (at
least if Conexant is the encoder, it looks like it is bypassed for other cases).
If there is a mismatch, the DirectX frame counter is incremented.

This can cause an issue where the vblank interrupt handler misses the expected
frame counter, causing it to fail to update the nv2a buffer read index via the
NV_PGRAPH_INCREMENT PGRAPH register. In this particular title, this leads to a
livelock where the pushbuffer is waiting on a buffer flip that cannot be
triggered until the current buffer is marked as read via the register write.

This change populates the low nibble of the Conexant device with a simple frame
counter. See https://xboxdevwiki.net/Video_Encoder for a link to the datasheet.

Unfortunately testing this on HW has proved difficult, as some state is changed
during D3D Present calls that causes SMBus reads to stop providing accurate
data. Thus, this change is mostly a guess at the actual behavior along with
some light testing without using D3D that shows the field value incrementing
from 0-3 and wrapping around forever, presumably aligned with vsync.

Fixes #1962