nv2a: Implement PTIMER alarm interrupts
Looks like the PTIMER alarm handling was started but never finished.
This implements my best guess at what the HW is doing (it's hard to test
with a nanosecond timer). It fixes the blocked loading for Def Jam: Fight
for NY, which I determined to be caused by a event handle never being set
due to the lack of an alarm interrupt in the same DPC that handles video
blanking and other PMC interrupts. The PTIMER pending flag is known from
the nxdk:
https://github.com/XboxDev/nxdk/blob/d0ae96acc926a6c3f6254b289a1a9a3c1db182f1/lib/pbkit/outer.h#L83

Fixes #1124